split up Stream infra
