+++ buffer/src/error.rs
@@ -56,7 +56,6 @@ impl Error {
     }
 
     pub(crate) fn invalid_value(reason: &'static str) -> Self {
-        panic!("{}", reason);
         Error(ErrorKind::InvalidValue { reason })
     }
