+++ buffer/src/value.rs
@@ -777,7 +777,6 @@ impl<'sval> sval::Stream<'sval> for ValueBuf<'sval> {
         label: Option<&sval::Label>,
         index: Option<&sval::Index>,
     ) -> sval::Result {
-        println!("enum_begin");
         #[cfg(feature = "alloc")]
         {
             self.push_begin(ValueKind::Enum {
@@ -802,8 +801,6 @@ impl<'sval> sval::Stream<'sval> for ValueBuf<'sval> {
         _: Option<&sval::Label>,
         _: Option<&sval::Index>,
     ) -> sval::Result {
-        println!("enum_end");
-
         #[cfg(feature = "alloc")]
         {
             self.try_catch(|buf| buf.push_end())
@@ -844,8 +841,6 @@ impl<'sval> sval::Stream<'sval> for ValueBuf<'sval> {
         _: Option<&sval::Label>,
         _: Option<&sval::Index>,
     ) -> sval::Result {
-        println!("tagged_end");
-
         #[cfg(feature = "alloc")]
         {
             self.try_catch(|buf| buf.push_end())
